Separate out the updates; improve documentation, reintroduce the MD5 legacy method for SD/HTTP based updating.
